uc-oms/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
python-dotenv>=1.0.0
cryptography>=42.0.0
pydantic>=2.6.0

